#25 split Netty3 and Netty4 to separate modules
* users need to add dependency either `replay-netty3` or `replay-netty4`
* each module has its own `play.server.Server` class
* replay own integration tests run on Netty4

implementation details:
* class `NettyInvocation` was not really needed anymore; now we just extend `Invocation`.
* I had to find a replacement for `Server.httpPort` which was only used in PDF module.
I hope `request.getBase()` is as good as `"http://localhost:" + Server.httpPort`. :)
