+++ build.gradle
@@ -63,6 +63,7 @@ subprojects {
 }
 
 ext {
+  commonsLangVersion = '2.6'
   slf4jVersion = '2.0.6'
   assertjVersion = '3.24.1'
   mockitoVersion = '4.11.0'
