+++ framework/build.gradle
@@ -14,7 +14,7 @@ dependencies {
   api('org.apache.commons:commons-email:1.5') {transitive = false}
   api('commons-fileupload:commons-fileupload:1.4')
   api('commons-io:commons-io:2.11.0')
-  implementation('commons-lang:commons-lang:2.6')
+  implementation("commons-lang:commons-lang:$commonsLangVersion")
   api('commons-logging:commons-logging:1.2')
   api('javax.mail:mail:1.4.7')
   api('javax.inject:javax.inject:1')
@@ -29,9 +29,6 @@ dependencies {
   api('org.hibernate:hibernate-validator:7.0.5.Final')
   api('javax.persistence:javax.persistence-api:2.2')
   implementation('org.hibernate:hibernate-ehcache:5.6.12.Final') {transitive = false}
-  // Upgrade to Netty4 is WIP: https://github.com/codeborne/replay/pull/25
-  api('io.netty:netty:3.10.6.Final')
-  api('io.netty:netty-all:4.1.87.Final')
   api("org.slf4j:slf4j-api:$slf4jVersion")
   api("org.slf4j:slf4j-reload4j:$slf4jVersion")
   api("org.slf4j:jul-to-slf4j:$slf4jVersion")
