+++ framework/src/play/Invoker.java
@@ -2,7 +2,6 @@ package play;
 
 import com.jamonapi.Monitor;
 import com.jamonapi.MonitorFactory;
-import play.server.NettyInvocation;
 import play.utils.PThreadFactory;
 
 import java.util.concurrent.Future;
@@ -32,7 +31,7 @@ public class Invoker {
      *            The code to run
      * @return The future object, to know when the task is completed
      */
-    public Future<?> invoke(NettyInvocation invocation) {
+    public Future<?> invoke(Invocation invocation) {
         Monitor monitor = MonitorFactory.getMonitor("Invoker queue size", "elmts.");
         monitor.add(executor.getQueue().size());
         invocation.onQueued();
