--- framework/src/play/server/NettyInvocation.java
@@ -1,6 +1,0 @@
-package play.server;
-
-import play.Invocation;
-
-public abstract class NettyInvocation extends Invocation {
-}
