--- framework/src/play/server/Server.java
@@ -1,13 +1,0 @@
-package play.server;
-
-import play.Play;
-
-public class Server extends play.server.netty4.Server {
-    public Server(Play play) {
-        super(play);
-    }
-
-    public Server(Play play, int port) {
-        super(play, port);
-    }
-}
