+++ netty3/src/play/server/netty3/FileChannelBuffer.java
@@ -6,7 +6,6 @@ import org.jboss.netty.buffer.ChannelBuffer;
 import org.jboss.netty.buffer.ChannelBufferFactory;
 import org.jboss.netty.buffer.ChannelBufferIndexFinder;
 import org.jboss.netty.buffer.WrappedChannelBuffer;
-import play.server.ResettableFileInputStream;
 
 import java.io.File;
 import java.io.FileNotFoundException;
