+++ netty3/src/play/server/netty3/ResettableFileInputStream.java
@@ -1,4 +1,4 @@
-package play.server;
+package play.server.netty3;
 
 import java.io.File;
 import java.io.FileInputStream;
