+++ netty3/test/play/server/netty3/ResettableFileInputStreamTest.java
@@ -1,4 +1,4 @@
-package play.server;
+package play.server.netty3;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
