+++ pdf/build.gradle
@@ -19,7 +19,7 @@ dependencies {
   testImplementation project(':framework').sourceSets.test.compileClasspath
   testImplementation("org.mockito:mockito-core:$mockitoVersion")
 
-  implementation('commons-lang:commons-lang:2.6')
+  implementation("commons-lang:commons-lang:$commonsLangVersion")
   thirdParty("org.xhtmlrenderer:flying-saucer-core:$FLYING_SOURCER_VERSION")
   implementation("org.xhtmlrenderer:flying-saucer-pdf:$FLYING_SOURCER_VERSION") {
     exclude group: 'org.xhtmlrenderer', module: 'flying-saucer-core'
