+++ settings.gradle
@@ -2,6 +2,8 @@ rootProject.name = 'replay'
 include 'framework'
 include 'guice'
 include 'liquibase'
+include 'netty3'
+include 'netty4'
 include 'fastergt'
 include 'pdf'
 include 'excel'
